c320: split Overview.tsx into 9 per-section files in pages/overview/ (#441 P0 2.2)
Overview.tsx 1074 LOC -> 56 LOC (-95%). Maintainability win, no bundle change
(sections are static imports, all visible on the landing page).
New files:
- pages/overview/types.ts (ScenePeek type + LABELLED_SCENES const)
- pages/overview/LandingCTA.tsx (50 LOC)
- pages/overview/HeroSpectralViz.tsx (262 LOC, includes HypercubeMini)
- pages/overview/HeadlineNumbers.tsx (51 LOC, includes HEADLINE_DEFS)
- pages/overview/FindingsCarousel.tsx (125 LOC, includes FINDINGS)
- pages/overview/HypercubeAnatomy.tsx (143 LOC)
- pages/overview/ScenesShowcase.tsx (124 LOC, includes SceneCard)
- pages/overview/PillarsTriptych.tsx (121 LOC, includes PillarIcon)
- pages/overview/MethodCoverage.tsx (78 LOC)
- pages/overview/ReadingPath.tsx (60 LOC)
Fixed import drift (useEffect, useState, useMemo, Link) in 4 section files.
typecheck clean, 44/44 tests pass.
